Add files via upload
This project is an implementation of RSA Algorithm, which is an Aysmmetric crptographic algorithm. it is known as the best public-key cryptosystem.
It is based on the difficulty of factoring large numbers.
in this project the implemented algorithm will go through three phases using two large prime numbers chosen by the user:
◦ Key generation
◦ Encryption 
◦ Decryption 
This project is not meant to be efficient it is just a challenge I liked to take.